refactor: writer 패키지 통합
